transfer types to file types.ts in urban telegram
